src/main/java/com/sora/springboot/web/HelloController.java
package com.sora.springboot.web;

import com.sora.springboot.web.dto.HelloResponseDto;
import org.springframework.web.bind.annotation.GetMapping;
import org.springframework.web.bind.annotation.RequestParam;
import org.springframework.web.bind.annotation.RestController;

/*
 * @RestController
 *      : JSON 데이터 타입으로 반환하는 컨트롤러로 지정해준다.
 *        각 메서드에서 @ResponseBody 선언한 것과 같다.
 */
@RestController
public class HelloController {

    @GetMapping("/hello")
    public String hello() {
        return "hello";
    }

    @GetMapping("/hello/dto")
    public HelloResponseDto helloDto(@RequestParam("name") String name, @RequestParam("amount") int amount) {
        return new HelloResponseDto(name, amount);
    }
}
